adding Complete form in /myForms

+++ S_and_C/src/main/java/com/example/s_and_c/Service/InternshipService.java
@@ -21,7 +21,7 @@ public interface InternshipService {
 
     void addFormResponse(FormResponseDTO internshipForStudentsDTO, String authEmail);
 
-    void addSelectedStudent(String email, int internshipId, String authEmail);
+    void addSelectedStudent(String email, long internshipId, String authEmail);
 
     void renounce(int internshipId, String authEmail);
 }
